requirements.txt
beautifulsoup4==4.13.4
botcity-framework-base==0.4.4
botcity-framework-core==1.3.0
botcity-framework-web==0.10.0
botcity-maestro-sdk==0.6.0
certifi==2025.4.26
charset-normalizer==3.4.2
comtypes==1.4.11
distro==1.9.0
idna==3.10
importlib_metadata==8.7.0
msedge-selenium-tools==3.141.3
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
psutil==7.0.0
pynput==1.8.1
pyperclip==1.9.0
python-dotenv==1.1.0
python-xlib==0.33
pywin32==310
pywinauto==0.6.9
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
selenium==3.141.0
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
urllib3==1.26.20
webdriver-manager==4.0.2
zipp==3.21.0
